[rename] hooks/gathering -> hooks/api, lib/api에 axios 인스턴스 및 비동기 에러처리 함수 이동
